feat: Command Model (쓰기용)과 Query Model (읽기용) 클래스를 분리 (#48)
