+++ docker-compose-batch.yml
@@ -1,0 +1,30 @@
+version: "3.8"
+
+services:
+  postgres:
+    image: postgres:15
+    container_name: jobterview-db
+    ports:
+      - "5432:5432"
+    environment:
+      POSTGRES_DB: jobterview
+      POSTGRES_USER: postgres
+      POSTGRES_PASSWORD: 1234
+    volumes:
+      - ./db:/var/lib/postgresql/data
+
+  rabbitmq:
+    image: rabbitmq:3-management
+    container_name: jobterview-rabbitmq
+    ports:
+      - "5672:5672"   # AMQP 기본 통신 포트
+      - "15672:15672" # 관리 콘솔 포트
+    volumes:
+      - ./rabbitmq_data:/var/lib/rabbitmq
+      - ./rabbitmq_log:/var/log/rabbitmq
+    environment:
+      - RABBITMQ_DEFAULT_USER=guest
+      - RABBITMQ_DEFAULT_PASS=guest
+    command: >
+      bash -c "rabbitmq-plugins enable rabbitmq_management
+      && rabbitmq-server"
