+++ docker-compose.yml
@@ -21,19 +21,3 @@ services:
     volumes:
       - ./redis.conf:/usr/local/etc/redis/redis.conf
     command: ["redis-server", "/usr/local/etc/redis/redis.conf"]
-
-  rabbitmq:
-    image: rabbitmq:3-management
-    container_name: jobterview-rabbitmq
-    ports:
-      - "5672:5672"   # AMQP 기본 통신 포트
-      - "15672:15672" # 관리 콘솔 포트
-    volumes:
-      - ./rabbitmq_data:/var/lib/rabbitmq
-      - ./rabbitmq_log:/var/log/rabbitmq
-    environment:
-      - RABBITMQ_DEFAULT_USER=guest
-      - RABBITMQ_DEFAULT_PASS=guest
-    command: >
-      bash -c "rabbitmq-plugins enable rabbitmq_management
-      && rabbitmq-server"
