+++ jobterview-api/src/main/kotlin/jobterview/api/job/controller/JobController.kt
@@ -2,8 +2,8 @@ package jobterview.api.job.controller
 
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
+import jobterview.api.job.query.service.usecase.JobQueryUseCase
 import jobterview.api.job.response.JobResponse
-import jobterview.api.job.service.JobService
 import jobterview.common.response.ApiResponse
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
@@ -13,12 +13,12 @@ import org.springframework.web.bind.annotation.RestController
 @RequestMapping("/jobs")
 @Tag(name = "Job", description = "직업 API")
 class JobController(
-    private val jobService: JobService
+    private val jobQueryUseCase: JobQueryUseCase
 ) {
 
     @Operation(summary = "직업 목록 조회")
     @GetMapping
     fun getJobs(): ApiResponse<List<JobResponse>> {
-        return ApiResponse.create(jobService.getJobs())
+        return ApiResponse.create(jobQueryUseCase.getJobs())
     }
 }
