+++ jobterview-api/src/main/kotlin/jobterview/api/job/query/service/usecase/JobQueryUseCase.kt
@@ -1,0 +1,11 @@
+package jobterview.api.job.query.service.usecase
+
+import jobterview.api.job.response.JobResponse
+import jobterview.domain.job.Job
+import java.util.*
+
+interface JobQueryUseCase {
+    fun getById(id: UUID): Job
+    fun getJobs(): List<JobResponse>
+    fun verifyExist(id: UUID)
+}
