+++ jobterview-api/src/main/kotlin/jobterview/api/question/controller/EmailQuestionController.kt
@@ -4,8 +4,8 @@ import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
 import jakarta.validation.constraints.Max
 import jakarta.validation.constraints.Min
+import jobterview.api.question.query.service.usecase.EmailQuestionQueryUseCase
 import jobterview.api.question.response.QuestionResponse
-import jobterview.api.question.service.EmailQuestionService
 import jobterview.api.question.vo.QuestionFilter
 import jobterview.common.response.ApiResponse
 import jobterview.domain.question.enums.QuestionDifficulty
@@ -22,7 +22,7 @@ import java.util.*
 @RequestMapping("/email-questions")
 @Tag(name = "Email Question")
 class EmailQuestionController (
-    private val emailQuestionService: EmailQuestionService
+    private val emailQuestionQueryUseCase: EmailQuestionQueryUseCase
 ) {
 
     @Operation(summary = "이메일별 전송된 질문 목록 조회")
@@ -48,6 +48,6 @@ class EmailQuestionController (
         val sortDirection = Sort.Direction.fromOptionalString(direction.uppercase()).orElse(Sort.Direction.DESC)
         val pageable = PageRequest.of(page, size, Sort.by(sortDirection, sort))
 
-        return ApiResponse.create(emailQuestionService.getQuestionsByEmail(email, token, filter, pageable))
+        return ApiResponse.create(emailQuestionQueryUseCase.getQuestionsByEmail(email, token, filter, pageable))
     }
 }
