+++ jobterview-api/src/main/kotlin/jobterview/api/subscription/controller/SubscriptionController.kt
@@ -2,38 +2,33 @@ package jobterview.api.subscription.controller
 
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
+import jobterview.api.subscription.command.service.usecase.SubscriptionCommandUseCase
+import jobterview.api.subscription.query.service.usecase.SubscriptionQueryUseCase
 import jobterview.api.subscription.request.SubscriptRequest
 import jobterview.api.subscription.request.VerifyEmailRequest
 import jobterview.api.subscription.response.SubscriptionResponse
-import jobterview.api.subscription.service.SubscriptionService
 import jobterview.common.response.ApiResponse
-import org.springframework.web.bind.annotation.DeleteMapping
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RequestParam
-import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.bind.annotation.*
 import java.util.*
 
 @RestController
 @RequestMapping("/subscriptions")
 @Tag(name = "Subscription", description = "구독 API")
 class SubscriptionController (
-    private val subscriptionService: SubscriptionService
+    private val subscriptionCommandUseCase: SubscriptionCommandUseCase,
+    private val subscriptionQueryUseCase: SubscriptionQueryUseCase,
 ){
 
     @Operation(summary = "구독 인증 메일 발송")
     @PostMapping("/verify/send")
     fun sendVerifyEmail(@RequestBody request: VerifyEmailRequest) {
-        subscriptionService.sendVerifyEmail(request)
+        subscriptionCommandUseCase.sendVerifyEmail(request)
     }
 
     @Operation(summary = "구독 요청")
     @PostMapping
     fun subscript(@RequestBody request: SubscriptRequest) {
-        subscriptionService.subscript(request)
+        subscriptionCommandUseCase.subscript(request)
     }
 
     @Operation(summary = "이메일별 구독 정보 목록 조회")
@@ -42,7 +37,7 @@ class SubscriptionController (
         @RequestParam email: String,
         @RequestParam token: String
     ): ApiResponse<List<SubscriptionResponse>> {
-        return ApiResponse.create(subscriptionService.getSubscriptions(email, token))
+        return ApiResponse.create(subscriptionQueryUseCase.getSubscriptionsByEmail(email, token))
     }
 
     @Operation(summary = "구독 해제")
@@ -52,6 +47,6 @@ class SubscriptionController (
         @RequestParam email: String,
         @RequestParam token: String
     ) {
-        subscriptionService.unsubscribe(id, email, token)
+        subscriptionCommandUseCase.unsubscribe(id, email, token)
     }
 }
