+++ jobterview-api/src/main/kotlin/jobterview/api/subscription/query/service/usecase/SubscriptionQueryUseCase.kt
@@ -1,0 +1,7 @@
+package jobterview.api.subscription.query.service.usecase
+
+import jobterview.api.subscription.response.SubscriptionResponse
+
+interface SubscriptionQueryUseCase {
+    fun getSubscriptionsByEmail(email: String, token: String): List<SubscriptionResponse>
+}
